Move filters Button in Call Hierarchy view to the top level Bar
- Added class FiltersAction
- Added fFiltersAction to CallHierarchyViewPart to create an icon in the
top level bar
- changed the constructor of FiltersAction so that the picture and Text
are the same as in the View Menu
- Connected the icon to its functionality by adding a method called
openFiltersDialog which opens the pop up window for the filters
